Major cleanup/refactor of layout code
