+++ apps/api/src/preferences.ts
@@ -1,6 +1,5 @@
 import {
   defaultCooccurrencePreferences,
-  defaultGraphLineRoutingStyle,
   defaultShowSingleFamilyTree,
   userPreferencesSchema,
   type CooccurrencePreferences,
@@ -26,7 +25,6 @@ export const withUserPreferenceDefaults = (
 
   return {
     ...parsed,
-    graphLineRoutingStyle: parsed.graphLineRoutingStyle ?? defaultGraphLineRoutingStyle,
     showSingleFamilyTree: parsed.showSingleFamilyTree ?? defaultShowSingleFamilyTree,
     primaryFamilyUnitByPersonId: parsed.primaryFamilyUnitByPersonId ?? {},
     cooccurrence: getCooccurrencePreferences(parsed)
@@ -41,7 +39,6 @@ export const mergeUserPreferences = (
   ...incoming,
   graphFilterVisibility: incoming.graphFilterVisibility ?? current.graphFilterVisibility,
   familyViewStyle: incoming.familyViewStyle ?? current.familyViewStyle,
-  graphLineRoutingStyle: incoming.graphLineRoutingStyle ?? current.graphLineRoutingStyle,
   showSingleFamilyTree: incoming.showSingleFamilyTree ?? current.showSingleFamilyTree,
   lastSelectedPersonId:
     incoming.lastSelectedPersonId !== undefined
