+++ apps/web/src/styles.css
@@ -164,6 +164,22 @@ body {
   padding: 12px;
 }
 
+.graph-layout-updating-banner {
+  position: absolute;
+  top: 12px;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 6;
+  padding: 6px 14px;
+  border-radius: 999px;
+  font-size: 0.875rem;
+  background: rgba(15, 23, 42, 0.88);
+  color: #e2e8f0;
+  border: 1px solid rgba(148, 163, 184, 0.35);
+  pointer-events: none;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25);
+}
+
 .hint {
   color: #cbd5e1;
   margin: 0;
