+++ packages/shared/src/index.ts
@@ -11,6 +11,15 @@ export const relationshipTypes = [
 ] as const;
 export type RelationshipType = (typeof relationshipTypes)[number];
 
+/** Relationship types that participate in family-tree layout topology (matches web `layout` + API graph layout). */
+export const graphLayoutTopologyRelationshipTypes = ["PARENT_OF", "CHILD_OF", "SPOUSE_OF"] as const;
+
+const graphLayoutTopologyTypeSet = new Set<RelationshipType>(graphLayoutTopologyRelationshipTypes);
+
+export const filterGraphLayoutTopologyRelationships = <R extends { type: RelationshipType }>(
+  relationships: R[]
+): R[] => relationships.filter((relationship) => graphLayoutTopologyTypeSet.has(relationship.type));
+
 export const genderValues = ["MALE", "FEMALE", "OTHER", "UNKNOWN"] as const;
 export type GenderValue = (typeof genderValues)[number];
 
@@ -153,17 +162,16 @@ export const graphFilterVisibilitySchema = z.object({
   pets: z.boolean()
 });
 
-export const familyViewStyleValues = [
+/** Legacy values accepted from stored preferences/API; normalized to generation tree only. */
+export const legacyFamilyViewStyleValues = [
   "generationTree",
   "centeredRelationshipMap",
   "hybridTreeList",
   "cleaned3D"
 ] as const;
-export const familyViewStyleSchema = z.enum(familyViewStyleValues);
-export const graphLineRoutingStyleValues = ["orthogonal", "direct"] as const;
-export const graphLineRoutingStyleSchema = z.enum(graphLineRoutingStyleValues);
-export type GraphLineRoutingStyle = z.infer<typeof graphLineRoutingStyleSchema>;
-export const defaultGraphLineRoutingStyle: GraphLineRoutingStyle = "orthogonal";
+const legacyFamilyViewStyleSchema = z.enum(legacyFamilyViewStyleValues);
+export const familyViewStyleSchema = legacyFamilyViewStyleSchema.transform(() => "generationTree" as const);
+export type FamilyViewStyle = "generationTree";
 export const defaultShowSingleFamilyTree = false;
 
 export const cooccurrencePreferencesSchema = z.object({
@@ -179,8 +187,9 @@ export const defaultCooccurrencePreferences: CooccurrencePreferences = {
 
 export const userPreferencesSchema = z.object({
   graphFilterVisibility: graphFilterVisibilitySchema.optional(),
-  familyViewStyle: familyViewStyleSchema.optional(),
-  graphLineRoutingStyle: graphLineRoutingStyleSchema.optional(),
+  familyViewStyle: legacyFamilyViewStyleSchema
+    .optional()
+    .transform((value) => (value === undefined ? undefined : ("generationTree" as const))),
   showSingleFamilyTree: z.boolean().optional(),
   lastSelectedPersonId: z.string().nullable().optional(),
   primaryFamilyUnitByPersonId: z.record(z.string(), z.string()).optional(),
@@ -189,7 +198,6 @@ export const userPreferencesSchema = z.object({
 });
 
 export type GraphFilterVisibility = z.infer<typeof graphFilterVisibilitySchema>;
-export type FamilyViewStyle = z.infer<typeof familyViewStyleSchema>;
 export type UserPreferences = z.infer<typeof userPreferencesSchema>;
 export type GraphLayoutMode = z.infer<typeof graphLayoutModeSchema>;
 
@@ -208,7 +216,9 @@ export const graphLayoutRequestSchema = z.object({
     })
   ),
   viewMode: graphLayoutModeSchema.default("family"),
-  familyViewStyle: familyViewStyleSchema.optional(),
+  familyViewStyle: legacyFamilyViewStyleSchema
+    .optional()
+    .transform((value) => (value === undefined ? undefined : ("generationTree" as const))),
   selectedPersonId: z.string().nullable().optional(),
   primaryFamilyUnitByPersonId: z.record(z.string(), z.string()).optional()
 });
@@ -232,12 +242,14 @@ const hashString = (input: string) => {
   return (hash >>> 0).toString(16).padStart(8, "0");
 };
 
+/** Revision for layout cache / server positions. Only parent/child/spouse edges affect the relationship hash (same as web layout topology). */
 export const buildGraphLayoutRevision = (request: GraphLayoutRequest) => {
   const peopleSignature = [...request.people]
     .sort((left, right) => left.id.localeCompare(right.id))
     .map((person) => `${person.id}:${person.name}`)
     .join(",");
-  const relationshipSignature = [...request.relationships]
+  const topologyRelationships = filterGraphLayoutTopologyRelationships(request.relationships);
+  const relationshipSignature = [...topologyRelationships]
     .sort((left, right) => {
       const leftKey = `${left.fromPersonId}|${left.type}|${left.toPersonId}`;
       const rightKey = `${right.fromPersonId}|${right.type}|${right.toPersonId}`;
@@ -253,7 +265,7 @@ export const buildGraphLayoutRevision = (request: GraphLayoutRequest) => {
     : "";
   return [
     `mode=${request.viewMode}`,
-    `style=${request.familyViewStyle ?? ""}`,
+    `style=${request.familyViewStyle ?? "generationTree"}`,
     `selected=${request.selectedPersonId ?? ""}`,
     `people=${hashString(peopleSignature)}`,
     `relationships=${hashString(relationshipSignature)}`,
